+++ RELEASE_NOTES.md
@@ -20,7 +20,7 @@ KeyboardKit 9 is planned to be released shortly after the public release of iOS 
 
 
 
-## 8.8.3
+## 8.9
 
 This version makes the emoji keyboard work in iOS 18.